requirements.txt
Flask
pdf2image
pytesseract
opencv-python
Pillow